Add change class with after, before, add change pom(change version webdrivermanager)
