+++ pom.xml
@@ -37,7 +37,7 @@
         <dependency>
             <groupId>io.github.bonigarcia</groupId>
             <artifactId>webdrivermanager</artifactId>
-            <version>4.4.3</version>
+            <version>5.4.1</version>
             <scope>test</scope>
         </dependency>
         <dependency>
