+++ src/main/java/ru/yandex/stellarburgers/LoginPage.java
@@ -1,4 +1,4 @@
-package stellarburgers;
+package ru.yandex.stellarburgers;
 import io.qameta.allure.Step;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -6,7 +6,6 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 public class LoginPage {
-
     private WebDriver driver;
     private By recoverPasswordButton = By.xpath(".//a[text()='Восстановить пароль']");
     private By registerButton = By.xpath(".//a[text()='Зарегистрироваться']");
@@ -17,28 +16,23 @@ public class LoginPage {
     public LoginPage(WebDriver driver) {
         this.driver = driver;
     }
-
     @Step("Заполнение формы входа")
     public void setLoginForm(String email, String password) {
         driver.findElement(emailInput).sendKeys(email);
         driver.findElement(passwordInput).sendKeys(password);
     }
-
     @Step("Нажатие кнопки зарегистрироваться")
     public void clickRegisterButton() {
         driver.findElement(registerButton).click();
     }
-
     @Step("Нажатие кнопки Войти")
     public void clickSingInButton() {
         driver.findElement(singInButton).click();
     }
-
     @Step("Нажатие кнопки восстановить пароль")
     public void clickRecoverPasswordButton() {
         driver.findElement(recoverPasswordButton).click();
     }
-
     @Step("Проверка отображения кнопки Войти")
     public boolean isSingInButtonDisplayed() {
         new WebDriverWait(driver, 20)
