+++ src/main/java/ru/yandex/stellarburgers/RegistrationPage.java
@@ -1,10 +1,9 @@
-package stellarburgers;
+package ru.yandex.stellarburgers;
 import io.qameta.allure.Step;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
 public class RegistrationPage {
-
     private WebDriver driver;
     private By nameFieldInput = By.xpath("//label[text()='Имя']//following-sibling::input");
     private By emailFieldInput = By.xpath("//label[text()='Email']//following-sibling::input");
@@ -16,27 +15,22 @@ public class RegistrationPage {
     public RegistrationPage(WebDriver driver) {
         this.driver = driver;
     }
-
     @Step("Заполнение поля Имя")
     public void setName(String name) {
         driver.findElement(nameFieldInput).sendKeys(name);
     }
-
     @Step("Заполнение поля Email")
     public void setEmail(String email) {
         driver.findElement(emailFieldInput).sendKeys(email);
     }
-
     @Step("Заполнение поля Пароль")
     public void setPassword(String password) {
         driver.findElement(passwordFieldInput).sendKeys(password);
     }
-
     @Step("Нажатие кнопки Зарегистрироваться")
     public void clickRegisterButton() {
         driver.findElement(registerButton).click();
     }
-
     @Step("Регистрация")
     public void setFieldsForRegistration(String name, String email, String password) {
         setName(name);
@@ -44,12 +38,10 @@ public class RegistrationPage {
         setPassword(password);
         clickRegisterButton();
     }
-
     @Step("Проверка отображения ошибки некорректного пароля")
     public boolean getTextError() {
         return driver.findElement(textError).isDisplayed();
     }
-
     @Step("Нажатие кнопки Войти")
     public void clickSingInButton() {
         driver.findElement(singInButton).click();
