+++ src/main/java/ru/yandex/stellarburgers/Rest.java
@@ -1,4 +1,4 @@
-package stellarburgers;
+package ru.yandex.stellarburgers;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
