+++ src/main/java/ru/yandex/stellarburgers/User.java
@@ -1,14 +1,12 @@
-package stellarburgers;
+package ru.yandex.stellarburgers;
 import io.qameta.allure.Step;
 import io.restassured.response.ValidatableResponse;
 import static io.restassured.RestAssured.given;
 
 public class User extends Rest {
-
     private static final String USER_REGISTER = "api/auth/register";
     private static final String USER_LOGIN = "api/auth/login";
     private static final String USER_DELETE = "api/auth/user";
-
     @Step("Регистрация пользователя")
     public ValidatableResponse createUser(UserMethods userMethods) {
         return given()
@@ -18,9 +16,7 @@ public class User extends Rest {
                 .post(USER_REGISTER)
                 .then()
                 .log().all();
-
     }
-
     @Step("Вход пользователя")
     public ValidatableResponse loginUser(UserCredentials credentials) {
         return given()
@@ -31,7 +27,6 @@ public class User extends Rest {
                 .then()
                 .log().all();
     }
-
     @Step("Удаление пользователя")
     public ValidatableResponse deleteUser(String accessToken) {
         return given()
