+++ src/main/java/ru/yandex/stellarburgers/UserCredentials.java
@@ -1,37 +1,22 @@
-package stellarburgers;
+package ru.yandex.stellarburgers;
 
-public class UserMethods {
-
-    private String name;
+public class UserCredentials {
     private String email;
     private String password;
 
-    public UserMethods(String name, String email, String password) {
-        this.name = name;
+    public UserCredentials(String email, String password) {
         this.email = email;
         this.password = password;
     }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
     public String getEmail() {
         return email;
     }
-
     public void setEmail(String email) {
         this.email = email;
     }
-
     public String getPassword() {
         return password;
     }
-
     public void setPassword(String password) {
         this.password = password;
     }
