+++ src/main/java/ru/yandex/stellarburgers/UserGenerator.java
@@ -1,15 +1,13 @@
-package stellarburgers;
+package ru.yandex.stellarburgers;
 import org.apache.commons.lang3.RandomStringUtils;
 
 public class UserGenerator {
-
     public static UserMethods getUser() {
         String name = RandomStringUtils.randomAlphabetic(8);
         String email = RandomStringUtils.randomAlphabetic(8) + "@new.ru";
         String password = RandomStringUtils.randomAlphabetic(8);
         return new UserMethods(name, email, password);
     }
-
     public static UserMethods getIncorrectPasswordUser() {
         String name = RandomStringUtils.randomAlphabetic(8);
         String email = RandomStringUtils.randomAlphabetic(8) + "@new.ru";
