+++ src/main/java/ru/yandex/stellarburgers/UserMethods.java
@@ -1,27 +1,28 @@
-package stellarburgers;
-
-public class UserCredentials {
-
+package ru.yandex.stellarburgers;
+public class UserMethods {
+    private String name;
     private String email;
     private String password;
-
-    public UserCredentials(String email, String password) {
+    public UserMethods(String name, String email, String password) {
+        this.name = name;
         this.email = email;
         this.password = password;
     }
-
+    public String getName() {
+        return name;
+    }
+    public void setName(String name) {
+        this.name = name;
+    }
     public String getEmail() {
         return email;
     }
-
     public void setEmail(String email) {
         this.email = email;
     }
-
     public String getPassword() {
         return password;
     }
-
     public void setPassword(String password) {
         this.password = password;
     }
