+++ src/test/java/AccountTests.java
@@ -1,70 +1,58 @@
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
-import stellarburgers.User;
-import stellarburgers.UserMethods;
-import stellarburgers.UserGenerator;
-import stellarburgers.AccountPage;
-import stellarburgers.LoginPage;
-import stellarburgers.MainPage;
+import ru.yandex.stellarburgers.User;
+import ru.yandex.stellarburgers.UserMethods;
+import ru.yandex.stellarburgers.UserGenerator;
+import ru.yandex.stellarburgers.AccountPage;
+import ru.yandex.stellarburgers.LoginPage;
+import ru.yandex.stellarburgers.MainPage;
 
 public class AccountTests extends Main {
-
     private User user;
     private UserMethods userMethods;
     private String accessToken;
-
     MainPage mainPage;
     LoginPage loginPage;
     AccountPage accountPage;
 
+    @Before
+    public void setUp() {
+        mainPage = new MainPage(webDriver);
+        mainPage.openPage();
+        mainPage.clickLoginButton();
+        user = new User();
+        userMethods = UserGenerator.getUser();
+        ValidatableResponse createResponse = user.createUser(userMethods);
+        accessToken = createResponse.extract().path("accessToken");
+        loginPage = new LoginPage(webDriver);
+        loginPage.setLoginForm(userMethods.getEmail(), userMethods.getPassword());
+        loginPage.clickSingInButton();
+        mainPage.clickPersonalAccountButton();
+        accountPage = new AccountPage(webDriver);
+    }
+    @After
+    public void deleteUser() {
+        user.deleteUser(accessToken);
+        System.out.println("Удален - " + accessToken);
+    }
     @Test
     @DisplayName("Переход по клику на <Личный кабинет>")
     public void loginUserPersonalAccountButton() {
-        mainPage = new MainPage(webDriver);
-        mainPage.openPage();
-        mainPage.clickPersonalAccountButton();
-        loginPage = new LoginPage(webDriver);
-        loginPage.isSingInButtonDisplayed();
+        accountPage.isLogOutButtonDisplayed();
     }
-
     @Test
     @DisplayName("Переход по клику на <Конструктор>")
     public void checkGoToConstructor() {
-        MainPage mainPage = new MainPage(webDriver);
-        mainPage.openPage();
-        mainPage.clickLoginButton();
-        user = new User();
-        userMethods = UserGenerator.getUser();
-        ValidatableResponse createResponse = user.createUser(userMethods);
-        accessToken = createResponse.extract().path("accessToken");
-        loginPage = new LoginPage(webDriver);
-        loginPage.setLoginForm(userMethods.getEmail(), userMethods.getPassword());
-        loginPage.clickSingInButton();
-        mainPage.clickPersonalAccountButton();
-        accountPage = new AccountPage(webDriver);
         accountPage.clickCreateBurgerButton();
         mainPage.isOrderButtonDisplayed();
-        user.deleteUser(accessToken);
     }
-
     @Test
-    @DisplayName("Переход по клику на логотипу Stellar Burgers")
+    @DisplayName("Переход по клику по логотипу Stellar Burgers")
     public void checkOutFromPersonalAccount() {
-        mainPage = new MainPage(webDriver);
-        mainPage.openPage();
-        mainPage.clickLoginButton();
-        user = new User();
-        userMethods = UserGenerator.getUser();
-        ValidatableResponse createResponse = user.createUser(userMethods);
-        accessToken = createResponse.extract().path("accessToken");
-        loginPage = new LoginPage(webDriver);
-        loginPage.setLoginForm(userMethods.getEmail(), userMethods.getPassword());
-        loginPage.clickSingInButton();
-        mainPage.clickPersonalAccountButton();
-        accountPage = new AccountPage(webDriver);
         accountPage.clickLogoStellarBurgerButton();
         mainPage.isOrderButtonDisplayed();
-        user.deleteUser(accessToken);
     }
 }
