+++ src/test/java/LoginTests.java
@@ -3,19 +3,17 @@ import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import stellarburgers.User;
-import stellarburgers.UserMethods;
-import stellarburgers.LoginPage;
-import stellarburgers.MainPage;
-import stellarburgers.RecoverPasswordPage;
-import stellarburgers.RegistrationPage;
-import stellarburgers.UserGenerator;
+import ru.yandex.stellarburgers.User;
+import ru.yandex.stellarburgers.UserMethods;
+import ru.yandex.stellarburgers.LoginPage;
+import ru.yandex.stellarburgers.MainPage;
+import ru.yandex.stellarburgers.RecoverPasswordPage;
+import ru.yandex.stellarburgers.RegistrationPage;
+import ru.yandex.stellarburgers.UserGenerator;
 
 public class LoginTests extends Main {
-
     private User user;
     private UserMethods userMethods;
-
     MainPage mainPage;
     LoginPage loginPage;
     RegistrationPage registerPage;
@@ -24,6 +22,11 @@ public class LoginTests extends Main {
     @Before
     public void setUp() {
         user = new User();
+        userMethods = UserGenerator.getUser();
+        ValidatableResponse createResponse = user.createUser(userMethods);
+        accessToken = createResponse.extract().path("accessToken");
+        mainPage = new MainPage(webDriver);
+        mainPage.openPage();
     }
     @After
     public void deleteUser() {
@@ -33,41 +36,24 @@ public class LoginTests extends Main {
     @Test
     @DisplayName("Вход через кнопку <Войти в аккаунт> на главной")
     public void loginUserLoginButton() {
-        userMethods = UserGenerator.getUser();
-        ValidatableResponse createResponse = user.createUser(userMethods);
-        accessToken = createResponse.extract().path("accessToken");
-        mainPage = new MainPage(webDriver);
-        mainPage.openPage();
         mainPage.clickLoginButton();
         loginPage = new LoginPage(webDriver);
         loginPage.setLoginForm(userMethods.getEmail(), userMethods.getPassword());
         loginPage.clickSingInButton();
         mainPage.isOrderButtonDisplayed();
     }
-
     @Test
     @DisplayName("Вход через кнопку <Личный кабинет>")
     public void loginUserPersonalAccountButton() {
-        userMethods = UserGenerator.getUser();
-        ValidatableResponse createResponse = user.createUser(userMethods);
-        accessToken = createResponse.extract().path("accessToken");
-        mainPage = new MainPage(webDriver);
-        mainPage.openPage();
         mainPage.clickPersonalAccountButton();
         loginPage = new LoginPage(webDriver);
         loginPage.setLoginForm(userMethods.getEmail(), userMethods.getPassword());
         loginPage.clickSingInButton();
         mainPage.isOrderButtonDisplayed();
     }
-
     @Test
     @DisplayName("Вход через кнопку в форме регистрации")
     public void loginUserRegistrationForm() {
-        userMethods = UserGenerator.getUser();
-        ValidatableResponse createResponse = user.createUser(userMethods);
-        accessToken = createResponse.extract().path("accessToken");
-        mainPage = new MainPage(webDriver);
-        mainPage.openPage();
         mainPage.clickLoginButton();
         loginPage = new LoginPage(webDriver);
         loginPage.clickRegisterButton();
@@ -77,15 +63,9 @@ public class LoginTests extends Main {
         loginPage.clickSingInButton();
         mainPage.isOrderButtonDisplayed();
     }
-
     @Test
     @DisplayName("Вход через кнопку в форме восстановления пароля")
     public void loginUserRecoverPasswordButton() {
-        userMethods = UserGenerator.getUser();
-        ValidatableResponse createResponse = user.createUser(userMethods);
-        accessToken = createResponse.extract().path("accessToken");
-        mainPage = new MainPage(webDriver);
-        mainPage.openPage();
         mainPage.clickLoginButton();
         loginPage = new LoginPage(webDriver);
         loginPage.clickRecoverPasswordButton();
