+++ src/test/java/Main.java
@@ -7,22 +7,17 @@ import org.openqa.selenium.chrome.ChromeOptions;
 import java.util.concurrent.TimeUnit;
 
 public class Main {
-
     WebDriver webDriver;
     ChromeOptions options;
-
     @Before
     public void init() {
-
         WebDriverManager.chromedriver().setup();
         webDriver = new ChromeDriver();
-
         //запуск в Yandex браузере
         /*options = new ChromeOptions();
         options.setBinary("/Applications/Yandex.app/Contents/MacOS/Yandex");
         WebDriverManager.chromedriver().setup();
         webDriver = new ChromeDriver(options);*/
-
         webDriver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
     }
     @After
